+++ spiffyplots/__init__.py
@@ -5,7 +5,7 @@ from pathlib import Path
 import matplotlib
 import matplotlib.style
 
-from . import cmap, colors
+from . import cmap, colors, journals
 from ._units import CM, MM, figsize
 from .lineplots import multiline
 from .multipanel import MultiPanel
@@ -23,6 +23,7 @@ STYLE_ALIASES = {
     "bright": "tol-bright",
     "muted": "tol-muted",
     "vibrant": "tol-vibrant",
+    **journals.ALIASES,
 }
 
 __all__ = [
@@ -34,6 +35,7 @@ __all__ = [
     "cmap",
     "colors",
     "figsize",
+    "journals",
     "label_panels",
     "multiline",
     "populate_random_data",